ubgrade end

+++ database/migrations/2021_01_10_133332_create_blog_categories_table.php
@@ -14,8 +14,8 @@ class CreateBlogCategoriesTable extends Migration
     public function up()
     {
         Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsiigned()->default(1);
+            $table->bigIncrements('id');
+            $table->bigInteger('parent_id')->unsiigned()->default(1);
             $table->string('slug')->unique();
             $table->string('title');
             $table->text('description')->nullable();
